[TESTS BROKEN] Massive multicursor rewrite...
The entire extension is rewritten to support multicursors. Due to
this rewrite, all tests are in a broken state. The tests will be fixed
in the next commit.

Changes:

1. The `Controller` class is completely rewritten. It now manages
`Tracker`s that each manage a cursor. This is the mechanism by which
we can have multicursor support.

2. The class previously known as `Pairs` is extensively optimised,
restructured then renamed to `Tracker`. Recall that `Pairs` was
previously used to track the pairs inserted by the primary cursor.
The new `Tracker` class has a similar function, but it is no longer
tied to only the primary cursor.

3. The `Settings` class is replaced with the `Configuration` class.
This new class was introduced in commit `8ec9ee6`. This change makes
the code more robust because it uses the `vscode-configuration-handler`
dependency that was introduced in commit 'acaf934'. Furthermore, unlike
`Settings`, the `Configuration` class will not use hardcoded default
values, and will instead rely on default values defined in the package
manifest.

4. The `Pair` class is rewritten to contain methods that were previously
free functions within `Pairs.ts`. Most importantly, the content change
application and the pair decoration routines are now embedded as methods
of `Pair`.

5. Registration of the active text editor and configuration change
watchers are moved out of `Controller` and into the main `activate()`
function to simplify the code structure.

6. Complementary to step 5, the `leap` and `escapeLeaperMode` command
registrations are moved from `Controller` into `activate()`.

7. Pairs are now selectively decorated: only pairs that are visible
within the viewport have decorations. This helps solve performance
issues when there are many multicursors beyond the range of the
viewport.

8. The `Pair` class and `shift()` function get their own files to
reduce clutter.

9. `shift()` is rewritten to improve performance and clarity.

10. Remove the `EXT_IDENT` global variable for the extension identifier.
We can do so because we no longer expect the extension's identifer to
change.

11. Expose the state of the controller to the tests via API exports.