Restructure tests after multicursor rewrite...
The previous commit bf7ed25 is a huge code rewrite which broke all the
tests. This commit seeks to update the tests to work with the changes
introduced.

Changes:

1. Rename what was previously known as 'integration tests' to 'single
cursor tests'. The tests are moved to a `cursor/single.test.ts` file.
It makes sense for us to make this change, as the 'integration tests'
previously only covered testing for a single cursor.

2. Significantly expand and reorganize a large part of the single
cursor tests. Tests are now organized into groups.

3. Also reorganize the shift function tests into groups.

4. Remove the central `extension.test.ts` file and instead run all
the single cursor and shift function tests within the respective
modules.

5. Number the test groups with Roman numerals instead of Arabic
numerals to reduce confusion, as the Mocha JS test runner labels
errors with Arabic numerals.

6. Move the preliminary tests to the single cursor test module. Run
them as the first test.

7. Move the integration test utility functions into `cursor/utilities.ts`.
These utility functions are now meant to be for the single cursor and
future multi cursor tests.

8. Update the utility functions in `cursor/utilities.ts` to work with
the rewritten extension.