Remove visibility based decoration... [cont.]
I don't think it is right for us to selectively decorate based on
visibility. There is some overhead involved in tracking visibility
states, and furthermore, it is difficult to test as the result is
dependent on the viewport size of vscode's testing instance.
Consider also that it would be difficult to test on headless vscode
instances such as when testing through command line via `npm test`,
since there is no "viewport" to speak of.

So instead, we leave it up to vscode to optimize decorations that are
outside the viewport.