Modify `leap` to move cursor without popping pairs... [cont.]
Previously we would pop the innermost pairs for each cursor whenever we
performed a leap. However this caused a problem when there were multiple
leap calls within a single event loop cycle. The first leap call would
remove pairs from the tracker but not update the `prevSortedCursors`
variable within the tracker. This caused the second leap call to be
rejected because the `hasLineOfSight` context recalculation would use
stale cursors.

Furthermore, because we popped pairs without synchronizing the tracker
to the cursors after the leap, incorrect keybinding context values were
calculated then broadcasted during end of cycle synchronizations. This
made keybinding toggling inconsistent.

A solution that was considered but ultimately rejected was calling
`syncToCursors` using the cursors at the end of the `leap` method.
The problem with the aforementioned approach was that calling
`syncToCursors` at the end of the `leap` method meant that from the
perspective of the tracker, there was no longer a sequential ordering
of information being passed to it.

Consider for instance a cycle where three consecutive leap calls are
made. When the cursors are moved during each call, selection change
events are fired. These selection change events are then received later
(typically within the next cycle) and passed to the tracker via
`syncToCursors`. Had we had called `syncToCursors` at the end of the
`leap` method, we would have (correctly) advanced the state of the
tracker only to have `syncToCursors` be called with cursor positions
"from the past" when the selection change events are finally received.

Our solution was to modify `leap` to move the cursors without removing
pairs from the tracker. This allowed for a graceful handling of multiple
leap calls within a single event loop, because all the leap calls will
end up setting the cursor to the same position, as the state of the
tracker (and therefore the innermost pairs) has not yet changed due to
not yet receiving new cursor positions.

With this commit, multiple consecutive leap calls are effectively merged
into one call, which allows for an unambiguous ordering of information
to be passed to the tracker. Keybinding toggling is made consistent
through this.

Note that since multiple leap calls are really only possible when the
`Tab` key (or whatever key is bound to the leap command) is held down,
the user does not notice the merging of calls, since it is unlikely the
user is able to count how many auto-repeated keypresses were made while
the key was held down.