Change 'column index' to 'character index'... [cont.]
All mentions of 'column index' are replaced with 'character index'.

We perform this change because 'character index' is a more accurate
description of the indices we are handling in the extension, since we
are often calculating the byte length of strings, which are always given
in terms of UTF-16 code units .

On the other hand, 'column index' is misleading because it implies
a correspondence with the column number shown in the bottom right of
vscode. The column number in the bottom right of vscode is in units of
physical monospace width, which means for instance, that an emoji
consisting of two or even four code units could only take up a single
unit width.

Obviously we are not using units of physical width within this extension
so it makes sense for us to change the name of the index to something
more appropriate.