Setup fresh test environment generation... [cont.]
Instead of using the `.test-environment` directory for tests, we copy
that directory to a temporary `.test-environment-tmp` that is then used
by the tests. The `.test-environment-tmp` directory is reset each time
tests are run.

Doing this means we don't risk permanently dirtying the test environment
when performing tests that involve changing configuration values in the
test workspace.