Remove the '-for-depr-config' workspaces in test env... [cont.]
They were added a few commits ago but are now found to be unnecessary.

The reason we no longer need them is that we are not able to read deprecated
configurations scoped to workspace folders, as those configurations could not
be set in any scope more specific than the workspace scope (unlike the newer
configurations).