Remove unnecessary set optimization... [cont.]
Previously, the array of visible text editors was converted to a set such
that the set of visible text editors could be queried for membership in
O(1) expected time. That was meant to prevent an O(n^2) runtime in the
loop that follows. However, that optimization was probably not necessary
since the number of visible text editors is often rather small anyways.
Plus, allocating a new set probably costs more than the time saved from
faster queries.