Hide 'Leap' and 'Escape Leaper Mode' from the command palette... [cont.]
The purpose of this change is to avoid cluttering the user's command
palette.

Note that this change is achieved in an indirect way by specifying "when"
context guards that require `editorTextFocus`. When the command palette
is open, the `editorTextFocus` context will be `false`, so neither command
will ever appear in the command palette. However, we have made sure to
fully specify the correct "when" contexts for both commands just in case
in the future vscode decides to have the `editorTextFocus` context be
`true` when the command palette is opened over a text editor.