Improve `rangeBehavior` exclusion schema... [cont.]
When the `rangeBehavior` property is specified by the user, the new
schema produces a clearer warning as it singles out the `rangeBehavior`
property as being the problem. The old schema on the other hand, would
highlight the entire object and give a vague warning that the object
"Matches a schema that is not allowed".