Fix crucial mistake when opening text editors... [cont.]
Previously, each test case was given a "provided" text editor that was
not in any of the workspace folders. However, an incorrect assumption was
made that the provided text editor did not inherit the configuration
values of any workspace folder. When the provided text editor (or any
untitled text document) is opened, it actually inherits the configuration
values of the workspace folder that was in focus at the time of the
opening. That caused a lot of problems because the configuration values of
the text editor provided to a test case would vary depending on which view
column was last in focus for the previous test case.

Thus, we have fixed the mistake by requiring each test case to explicitly
open a file from the test workspace instead of being provided one. Now it
is much clearer which file inherits what configuration value.