Add step to close all tabs when test instance starts... [cont.]
This step is needed because the previous commit moved the test workspace
cleanup steps back into the post-test cleanup method. If we don't add
this step, then the first test in a test run might spuriously fail since
it might start with the editor in a view state brought over from a
previous failed run.