Allow configuration values to be hot reloaded... [cont.]
It was realized that clearing pairs after a change in configuration
was unnecessary. Instead, when options pertaining to decorations are
changed, all that needs to be done is to reapply the decorations. And
if  `detectedPairs` has changed, then it is safe to just use the new
value from that point on and keep the previously detected pairs (if any)
as they were. The tracking logic is "character blind", in that it does
not care about which characters make up the pair once a pair has been
detected. Thus, hot reloading the `detectedPairs` configuration is safe.