Revert to using decoration timer instead of queue... [cont.]
Using a queue caused some issues because it was possible for pairs that
have been added to the queue to be dropped by the tracker before the
decoration has been applied. That meant that the queue was decorating pairs
that have been dropped.

Thus, we have reverted to applying decorations by iterating through the
cluster. This allows us to get the most up-to-date state of the pairs when
applying decorations.