Decorate on-demand instead of delaying them... [cont.]
Previously, a delay was used to apply decorations at a later time when
hopefully there was no content change that could have caused the
decoration request to refer to a stale position. However, that did not
really solve the problem because we could not actually predict when the
next content change was going to come, so a delay of 30ms for instance,
was able to prevent content changes occurring at the ~10ms or ~20ms mark
from messing up the decoration, but was not able to do so for content
changes occurring right about the ~30ms mark.

Instead, we have devised this new approach where we decorate a new pair
right as it is detected and record down the time when the decoration was
made. Then, if any content changes occur within a short period of time
(which we call the redecoration window) after a decoration was applied,
we reapply said decoration. Through this approach, we were able to mostly
eliminate the problem of stale positions being decorated. In cases where
there is much lag in vscode however, stale positions can still be decorated
due to vscode only applying decorations long after the redecoration window;
but those situations are rare.